requirements.txt
alabaster==0.7.7
alembic==0.8.7
anaconda-client==1.4.0
anaconda-navigator==1.1.0
aniso8601==1.1.0
argcomplete==1.0.0
astropy==1.1.2
attrs==16.3.0
autopep8==1.2.4
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.9.1
bleach==1.4.2
blinker==1.4
bokeh==0.11.1
boto==2.39.0
Bottleneck==1.0.0
bz2file==0.98
cdecimal==2.3
cffi==1.5.2
chest==0.2.3
cloudpickle==0.1.1
clyent==1.2.1
colorama==0.3.7
comtypes==1.1.2
conda==4.0.5
conda-build==1.20.0
conda-env==2.4.5
conda-manager==0.3.1
configobj==5.0.6
constantly==15.1.0
cryptography==1.3
cssselect==1.0.0
cycler==0.10.0
Cython==0.23.4
cytoolz==0.7.5
dask==0.8.1
datashape==0.5.1
decorator==4.0.9
dill==0.2.4
Django==1.10.3
django-bootstrap3==7.0.1
django-el-pagination==3.0.1
django-model-utils==2.5.2
django-notifications-hq==1.2
docutils==0.12
dominate==2.2.1
echarts-python==0.1.2
enum34==1.1.2
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Bootstrap==3.3.5.7
Flask-Cors==2.1.2
Flask-HTTPAuth==2.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Moment==0.5.1
flask-mongoengine==0.7.5
Flask-OpenID==1.2.5
Flask-PageDown==0.2.1
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flup==1.0.2
funcsigs==0.4
future==0.15.2
futures==3.0.3
gensim==0.13.3
gevent==1.1.0
greenlet==0.4.9
grin==1.2.1
h5py==2.5.0
HeapDict==1.0.0
html5lib==0.999999999
htmlPy==2.0.3
idna==2.0
incremental==16.10.1
ipaddress==1.0.14
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
jieba==0.38
Jinja2==2.8
jsonfield==1.0.3
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
llvmlite==0.9.0
locket==0.2.0
logging==0.4.9.6
lxml==3.6.0
Mako==1.0.4
Markdown==2.6.5
MarkupSafe==0.23
matplotlib==1.5.1
menuinst==1.3.2
mistune==0.7.2
mongoengine==0.10.6
mpmath==0.19
multipledispatch==0.4.8
MySQL-python==1.2.5
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.2
nose==1.3.7
notebook==4.1.0
numba==0.24.0
numexpr==2.5
numpy==1.10.4
odo==0.4.2
openpyxl==2.3.2
packaging==16.7
pandas==0.18.0
parsel==1.1.0
partd==0.3.2
path.py==0.0.0
patsy==0.4.0
pefile==2016.3.28
pep8==1.7.0
pickleshare==0.5
Pillow==3.1.1
ply==3.8
psutil==4.1.0
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
PyDispatcher==2.0.5
pydot==1.2.2
pyflakes==1.1.0
Pygments==2.1.1
pygraphviz==1.3.1
PyInstaller==3.3.dev0+b78bfe5
pymongo==3.3.0
PyMySQL==0.7.2
pyOpenSSL==0.15.1
pyparsing==2.1.8
pyreadline==2.1
PySide==1.2.4
pytest==2.8.5
python-dateutil==2.5.1
python-editor==1.0.1
python-openid==2.2.5
pytz==2016.2
pywin32==220
PyYAML==3.11
pyzmq==15.2.0
QtAwesome==0.3.2
qtconsole==4.2.0
QtPy==1.0
queuelib==1.4.2
reportlab==3.3.0
requests==2.9.1
rope==0.9.4
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.17.0
Scrapy==1.3.0
scripts==2.0
seaborn==0.7.1
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.3.5
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
sphinx==1.3.5
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.0.12
statsmodels==0.6.1
stua==0.2
sympy==1.0
tables==3.2.2
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
Twisted==16.6.0
unicodecsv==0.14.1
visitor==0.1.3
w3lib==1.16.0
webencodings==0.5
Werkzeug==0.11.4
Whoosh==2.7.4
WTForms==2.1
xlrd==0.9.4
XlsxWriter==0.8.4
xlwings==0.7.0
xlwt==1.0.0
zope.interface==4.3.3
